Refactored MobileNavigation into a separate component to improve componentization, separating it from the Header componen
